Feature/mypage test

+++ src/main/java/com/github/yumyum/auth/dto/LogoutReqDto.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class LogoutReqDto {
 
-    private Long memberId;
+    private Integer memberId;
 
 }
